+++ Changes.md
@@ -14,6 +14,7 @@ Improvements
 - SceneInspector : Improved search filter widget.
 - AttributeEditor, HierarchyView, RenderPassEditor, SceneInspector, SetEditor : Improved search filter widgets.
 - Spreadsheet : Clicking on the search icon now moves the keyboard focus to the search field ready for editing.
+- InteractiveRender : Added `useVisibleSet` plug. When on, only the scene locations contained in the Visible Set will be rendered.
 
 Fixes
 -----